Only kill the old server if it exists

+++ run.sh
@@ -4,5 +4,7 @@ webpack
 cargo build --release
 
 OLD_GROUP_ID=$(ps x -o  "%p %r %y %x %c " | grep cargo | awk  '{print $2}')
-kill -- -$OLD_GROUP_ID
+if [ -n $OLD_GROUP_ID ]; then
+	kill -- -$OLD_GROUP_ID
+fi
 RUST_LOG=info cargo run --release &> output
